HW2/requirements.txt
aiogram==3.17.0
aiohttp==3.11.12
requests==2.32.3
matplotlib==3.9.4
numpy==2.0.2
python-dotenv==1.0.1
emoji==2.14.1